DPP-131 implementation of unnesting - review fixes - using Conversions for implicits
