+++ ledger/participant-integration-api/src/main/scala/platform/store/Conversions.scala
@@ -3,7 +3,7 @@
 
 package com.daml.platform.store
 
-import java.sql.PreparedStatement
+import java.sql.{PreparedStatement, Timestamp}
 import java.time.Instant
 import java.util.Date
 
@@ -191,6 +191,38 @@ private[platform] object Conversions {
   implicit object StringArrayParameterMetadata extends ParameterMetaData[Array[String]] {
     override def sqlType: String = "ARRAY"
     override def jdbcType: Int = java.sql.Types.ARRAY
+  }
+
+  implicit object ByteArrayArrayToStatement extends ToStatement[Array[Array[Byte]]] {
+    override def set(s: PreparedStatement, index: Int, v: Array[Array[Byte]]): Unit = {
+      val conn = s.getConnection
+      val ts = conn.createArrayOf("BYTEA", v.asInstanceOf[Array[AnyRef]])
+      s.setArray(index, ts)
+    }
+  }
+
+  implicit object CharArrayToStatement extends ToStatement[Array[String]] {
+    override def set(s: PreparedStatement, index: Int, v: Array[String]): Unit = {
+      val conn = s.getConnection
+      val ts = conn.createArrayOf("VARCHAR", v.asInstanceOf[Array[AnyRef]])
+      s.setArray(index, ts)
+    }
+  }
+
+  implicit object TimestampArrayToStatement extends ToStatement[Array[Timestamp]] {
+    override def set(s: PreparedStatement, index: Int, v: Array[Timestamp]): Unit = {
+      val conn = s.getConnection
+      val ts = conn.createArrayOf("TIMESTAMP", v.asInstanceOf[Array[AnyRef]])
+      s.setArray(index, ts)
+    }
+  }
+
+  implicit object InstantArrayToStatement extends ToStatement[Array[Instant]] {
+    override def set(s: PreparedStatement, index: Int, v: Array[Instant]): Unit = {
+      val conn = s.getConnection
+      val ts = conn.createArrayOf("TIMESTAMP", v.map(java.sql.Timestamp.from))
+      s.setArray(index, ts)
+    }
   }
 
 }
