+++ ledger/participant-integration-api/src/main/scala/platform/store/dao/events/ContractWitnessesTablePostgres.scala
@@ -9,7 +9,6 @@ import anorm.{Row, SimpleSql, SqlQuery}
 import com.daml.platform.store.dao.events.ContractWitnessesTable.Executable
 
 object ContractWitnessesTablePostgres extends ContractWitnessesTable {
-  import AnormParamsMapper._
 
   val witnessesContractIdsParam = "witnessesContractIds"
   val partiesParam = "parties"
@@ -31,6 +30,7 @@ object ContractWitnessesTablePostgres extends ContractWitnessesTable {
   private def buildInsertExecutable(
       witnesses: TransactionIndexing.ContractWitnessesInfo
   ): Executable = {
+    import com.daml.platform.store.Conversions._
 
     val flattened: Iterator[(ContractId, String)] = Relation.flatten(witnesses.netVisibility)
     val (witnessesContractIds, parties) = flattened
